pyproject.toml
[project]
name = "speechall-python-example"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "speechall==0.2.0",
    "gradio==3.50.0",
    "openai>=1.98.0",
    "python-dotenv>=1.0.0",
    "audioop-lts>=0.2.1",
    "setuptools>=80.9.0",
    "wheel>=0.45.1",
]
